[wndx_sane_tgt_rm_opt.cmake] This function does not work!
NOTE: this module has not been tested properly, it does not currently work and may never work!
The idea of removing options/flags from the targets might be a dead end.
The only correct way is not to inherit options/flags that are not needed!

This file is a reminder that experiments were done with this feature.